gabung cuci

+++ admin/transaksi.php
@@ -13,8 +13,7 @@ $dataTransaksi = $clientTransaksi->tampil_semua_transaksi();
     <title>Data Transaksi</title>
 
     <!-- Google Font: Source Sans Pro -->
-    <link rel="stylesheet"
-        href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700&display=fallback">
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700&display=fallback">
     <!-- Font Awesome -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css">
     <!-- DataTables -->
@@ -33,194 +32,17 @@ $dataTransaksi = $clientTransaksi->tampil_semua_transaksi();
         </div>
 
         <!-- Navbar -->
-        <nav class="main-header navbar navbar-expand navbar-white navbar-light">
-            <!-- Left navbar links -->
-            <ul class="navbar-nav">
-                <li class="nav-item">
-                    <a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
-                </li>
-                <li class="nav-item d-none d-sm-inline-block">
-                    <a href="./index.php" class="nav-link">Home</a>
-                </li>
-                <li class="nav-item d-none d-sm-inline-block">
-                    <a href="#" class="nav-link">Contact</a>
-                </li>
-            </ul>
-        </nav>
+        <?php
+        include "navbar.php";
+        ?>
         <!-- /.navbar -->
 
-        <!-- Main Sidebar Container -->
-        <aside class="main-sidebar sidebar-dark-primary elevation-4">
-            <!-- Brand Logo -->
-            <a href="./index.php" class="brand-link">
-                <img src="./dist/img/hishoes.png" alt="AdminLTE Logo" class="brand-image img-circle elevation-3"
-                    style="opacity: .8">
-                <span class="brand-text font-weight-light">HISHOES</span>
-            </a>
-
-            <!-- Sidebar -->
-            <div class="sidebar">
-                <!-- Sidebar user panel (optional) -->
-                <div class="user-panel mt-3 pb-3 mb-3 d-flex">
-                    <div class="image">
-                        <img src="./dist/img/hishoes-160x160.png" class="img-circle elevation-2" alt="User Image">
-                    </div>
-                    <div class="info">
-                        <p class="text-white">ADMIN</p>
-                    </div>
-                </div>
-
-                <!-- Sidebar Menu -->
-                <nav class="mt-2">
-                    <ul class="nav nav-pills nav-sidebar flex-column" data-widget="treeview" role="menu"
-                        data-accordion="false">
-                        <!-- Add icons to the links using the .nav-icon class
-                 with font-awesome or any other icon font library -->
-                        <li class="nav-item">
-                            <a href="./index.php" class="nav-link ">
-                                <i class="nav-icon fas fa-tachometer-alt"></i>
-                                <p>
-                                    Dashboard
-                                </p>
-                            </a>
-                        </li>
-
-                        <!-- Produk -->
-                        <li class="nav-header">Produk</li>
-                        <li class="nav-item">
-                            <a href="#" class="nav-link">
-                                <i class="fas fa-shoe-prints"></i>
-                                <p>
-                                    Produk
-                                    <i class="fas fa-angle-left right"></i>
-                                </p>
-                            </a>
-                            <ul class="nav nav-treeview">
-                                <li class="nav-item">
-                                    <a href="./produk.php" class="nav-link">
-                                        <i class="far fa-dot-circle nav-icon"></i>
-                                        <p>Data produk</p>
-                                    </a>
-                                </li>
-                                <li class="nav-item">
-                                    <a href="./form-produk.php" class="nav-link">
-                                        <i class="far fa-plus-square nav-icon"></i>
-                                        <p>Tambah produk</p>
-                                    </a>
-                                </li>
-                            </ul>
-                        </li>
-                        <li class="nav-item">
-                            <a href="#" class="nav-link">
-                                <i class="nav-icon fas fa-digital-tachograph"></i>
-                                <p>
-                                    Detail Produk
-                                    <i class="fas fa-angle-left right"></i>
-                                </p>
-                            </a>
-                            <ul class="nav nav-treeview">
-                                <li class="nav-item">
-                                    <a href="./detail-produk.php" class="nav-link">
-                                        <i class="far fa-dot-circle nav-icon"></i>
-                                        <p>Data Detail Produk</p>
-                                    </a>
-                                </li>
-                                <li class="nav-item">
-                                    <a href="./form-detail-produk.php" class="nav-link">
-                                        <i class="far fa-plus-square nav-icon"></i>
-                                        <p>Tambah Detail Produk</p>
-                                    </a>
-                                </li>
-                            </ul>
-                        </li>
-                        <li class="nav-item">
-                            <a href="#" class="nav-link">
-                                <i class="fas fa-poll-h"></i>
-                                <p>
-                                    Kategori
-                                    <i class="fas fa-angle-left right"></i>
-                                </p>
-                            </a>
-                            <ul class="nav nav-treeview">
-                                <li class="nav-item">
-                                    <a href="./kategori.php" class="nav-link">
-                                        <i class="far fa-dot-circle nav-icon"></i>
-                                        <p>Data Kategori Produk</p>
-                                    </a>
-                                </li>
-                                <li class="nav-item">
-                                    <a href="./form-kategori.php" class="nav-link">
-                                        <i class="far fa-plus-square nav-icon"></i>
-                                        <p>Tambah Kategori Produk</p>
-                                    </a>
-                                </li>
-                            </ul>
-                        </li>
-
-
-                        <!-- PELANGGAN -->
-                        <li class="nav-header">Pelangan</li>
-                        <li class="nav-item">
-                            <a href="#" class="nav-link ">
-                                <i class="nav-icon fas fa-id-badge"></i>
-                                <p>
-                                    Pelanggan
-                                    <i class="fas fa-angle-left right"></i>
-                                </p>
-                            </a>
-                            <ul class="nav nav-treeview">
-                                <li class="nav-item">
-                                    <a href="./pelanggan.php" class="nav-link">
-                                        <i class="far fa-dot-circle nav-icon"></i>
-                                        <p>Data Pelanggan</p>
-                                    </a>
-                                </li>
-                                <li class="nav-item">
-                                    <a href="./form-pelanggan.php" class="nav-link">
-                                        <i class="far fa-plus-square nav-icon"></i>
-                                        <p>Tambah Pelanggan</p>
-                                    </a>
-                                </li>
-                            </ul>
-                        </li>
-
-                        <!-- TRANSAKSI -->
-                        <li class="nav-header">Transaksi</li>
-                        <li class="nav-item">
-                            <a href="#" class="nav-link active">
-                                <i class="fas fa-cart-arrow-down"></i>
-                                <p>
-                                    Transaksi
-                                    <i class="fas fa-angle-left right"></i>
-                                    <span class="badge badge-success right">
-                                        <?php echo count($dataTransaksi) ?>
-                                    </span>
-                                </p>
-                            </a>
-                            <ul class="nav nav-treeview">
-                                <li class="nav-item">
-                                    <a href="./transaksi.php" class="nav-link">
-                                        <i class="far fa-dot-circle nav-icon"></i>
-                                        <p>Data Transaksi</p>
-                                    </a>
-                                </li>
-                                <li class="nav-item">
-                                    <a href="./form-transaksi.php" class="nav-link">
-                                        <i class="far fa-plus-square nav-icon"></i>
-                                        <p>Tambah Transaksi</p>
-                                    </a>
-                                </li>
-                            </ul>
-                        </li>
-
-                    </ul>
-                    </li>
-                    </ul>
-                </nav>
-                <!-- /.sidebar-menu -->
-            </div>
-            <!-- /.sidebar -->
-        </aside>
+        <!-- Sidebar -->
+        <?php
+        $activePage = 'transaksi';
+        include "sidebar.php";
+        ?>
+        <!-- /.sidebar -->
 
         <!-- Content Wrapper. Contains page content -->
         <div class="content-wrapper">
@@ -229,12 +51,12 @@ $dataTransaksi = $clientTransaksi->tampil_semua_transaksi();
                 <div class="container-fluid">
                     <div class="row mb-2">
                         <div class="col-sm-6">
-                            <h1>Tambah Transaksi</h1>
+                            <h1>Transaksi</h1>
                         </div>
                         <div class="col-sm-6">
                             <ol class="breadcrumb float-sm-right">
                                 <li class="breadcrumb-item"><a href="./index.php">Home</a></li>
-                                <li class="breadcrumb-item active">Tambah Transaksi</li>
+                                <li class="breadcrumb-item active"> Transaksi</li>
                             </ol>
                         </div>
                     </div>
@@ -268,7 +90,7 @@ $dataTransaksi = $clientTransaksi->tampil_semua_transaksi();
                                             <?php $no = 1;
                                             $data_array = $clientTransaksi->tampil_semua_transaksi();
                                             foreach ($data_array as $r) {
-                                                ?>
+                                            ?>
                                                 <tr>
                                                     <td>
                                                         <?= $no ?>
@@ -291,16 +113,11 @@ $dataTransaksi = $clientTransaksi->tampil_semua_transaksi();
                                                         <?= $r->jumlah ?>
                                                     </td>
                                                     <td class="col-2 text-center">
-                                                        <!-- <a class="btn btn-outline-primary btn-sm" type="button" a
-                                                                href="ubah-transaksi.php?id_transaksi=<?= $r->id_transaksi ?>"><i class="fas fa-edit"></i> Ubah</a> -->
-                                                        <a class="btn btn-outline-danger btn-sm" type="button" a
-                                                            href="proses-transaksi.php?aksi=hapus&id_transaksi=<?= $r->id_transaksi ?>"
-                                                            onclick="return confirm('Apakah Anda Ingin Menghapus Data Ini?')"><i
-                                                                class="fas fa-trash-alt"></i> Hapus</a>
+                                                        <a class="btn btn-outline-danger btn-sm" type="button" a href="proses-transaksi.php?aksi=hapus&id_transaksi=<?= $r->id_transaksi ?>" onclick="return confirm('Apakah Anda Ingin Menghapus Data Ini?')"><i class="fas fa-trash-alt"></i> Hapus</a>
                                                     </td>
 
                                                 </tr>
-                                                <?php $no++;
+                                            <?php $no++;
                                             }
                                             unset($data_array, $r, $no);
                                             ?>
@@ -322,14 +139,10 @@ $dataTransaksi = $clientTransaksi->tampil_semua_transaksi();
             <!-- /.content -->
         </div>
         <!-- /.content-wrapper -->
-        <footer class="main-footer">
-            <strong>Copyright &copy; 2023 <a href="https://instagram.com/henyrmdn_" target="_blank">Heny Rimadana</a> &
-                <a href="https://instagram.com/imamtl.k" target="_blank">Imamatul Khoiriyah</a>.</strong>
-            All rights reserved.
-            <div class="float-right d-none d-sm-inline-block">
-                <b>Sistem Terdistribusi & Keamanan</b> 2023
-            </div>
-        </footer>
+
+        <!-- Footer -->
+        <?php include "footer.php" ?>
+        <!-- Footer -->
 
         <!-- Control Sidebar -->
         <aside class="control-sidebar control-sidebar-dark">
